pyproject.toml
[tool.poetry]
name = "ibge-geodata"
version = "2.0.0"
description = "Python package to use spatial data of Brazilian territorial levels available at IBGE"
authors = ["Victor Benezoli <victorbenezoli@gmail.com>"]
license = "GPL"
readme = "README.md"
repository = "https://github.com/victorbenezoli/geodata"
homepage = "https://github.com/victorbenezoli/geodata"
packages = [
    { include = "geodata", from = "." },
]
include = ["geodata/**/*.py", "README.md"]
exclude = ["tests/*", "docs/*"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License (GPL)",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.11"
geopandas = "^1.0.1"
requests = "^2.32.3"
pytest = "^8.3.4"
numpy = "^2.2.3"
rich = "^14.0.0"
matplotlib = "^3.10.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "*"
mkdocs-minify-plugin = "*"
pymdown-extensions = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
